+++ wsidicomizer/extras/bioformats/bioformats_image_data.py
@@ -22,7 +22,7 @@ from pydicom.uid import UID
 from wsidicom.codec import Encoder
 from wsidicom.geometry import Point, Region, Size, SizeMm
 
-from wsidicomizer.config import settings
+from wsidicomizer.config import get_settings
 from wsidicomizer.extras.bioformats.bioformats_reader import BioformatsReader
 from wsidicomizer.image_data import BaseDicomizerImageData
 
@@ -39,7 +39,7 @@ class BioformatsImageData(BaseDicomizerImageData):
     ) -> None:
         super().__init__(encoder)
         if tile_size is None:
-            tile_size = settings.default_tile_size
+            tile_size = get_settings().default_tile_size
         self._tile_size = Size(tile_size, tile_size)
         self._image_reader = reader
         self._image_index = image_index